+++ libindy/src/services/crypto/mod.rs
@@ -49,7 +49,7 @@ impl CryptoService {
     }
 
     pub fn create_key(&self, key_info: &KeyInfo) -> Result<Key, CryptoError> {
-        trace!("create_key >>> key_info: {:?}", key_info);
+        trace!("create_key >>> key_info: {:?}", secret!(key_info));
 
         let crypto_type_name = key_info.crypto_type
             .as_ref()
@@ -81,7 +81,7 @@ impl CryptoService {
     }
 
     pub fn create_my_did(&self, my_did_info: &MyDidInfo) -> Result<(Did, Key), CryptoError> {
-        trace!("create_my_did >>> my_did_info: {:?}", my_did_info);
+        trace!("create_my_did >>> my_did_info: {:?}", secret!(my_did_info));
 
         let crypto_type_name = my_did_info.crypto_type
             .as_ref()
@@ -296,6 +296,36 @@ impl CryptoService {
         Ok(decrypted_doc)
     }
 
+    pub fn authenticated_encrypt(&self, my_key: &Key, their_vk: &str, doc: &[u8]) -> Result<Vec<u8>, CryptoError> {
+        let msg = self.create_combo_box(my_key, their_vk, doc)?;
+
+        let msg_pack = msg.to_msg_pack()
+            .map_err(|e| CommonError::InvalidState(format!("Can't serialize ComboBox: {:?}", e)))?;
+
+        let res = self.encrypt_sealed(&their_vk, &msg_pack)?;
+
+        Ok(res)
+    }
+
+    pub fn authenticated_decrypt(&self, my_key: &Key, doc: &[u8]) -> Result<(String, Vec<u8>), CryptoError> {
+        let decrypted_msg = self.decrypt_sealed(&my_key, &doc)?;
+
+        let parsed_msg = ComboBox::from_msg_pack(decrypted_msg.as_slice())
+            .map_err(|err| CommonError::InvalidStructure(format!("Can't deserialize ComboBox: {:?}", err)))?;
+
+        let doc: Vec<u8> = base64::decode(&parsed_msg.msg)
+            .map_err(|err| CommonError::InvalidStructure(format!("Can't decode internal msg filed from base64 {}", err)))?;
+
+        let nonce: Vec<u8> = base64::decode(&parsed_msg.nonce)
+            .map_err(|err| CommonError::InvalidStructure(format!("Can't decode nonce from base64 {}", err)))?;
+
+        let decrypted_msg = self.decrypt(&my_key, &parsed_msg.sender, &doc, &nonce)?;
+
+        let res = (parsed_msg.sender, decrypted_msg);
+
+        Ok(res)
+    }
+
     pub fn encrypt_sealed(&self, their_vk: &str, doc: &[u8]) -> Result<Vec<u8>, CryptoError> {
         trace!("encrypt_sealed >>> their_vk: {:?}, doc: {:?}", their_vk, doc);
 
@@ -349,7 +379,7 @@ impl CryptoService {
     }
 
     pub fn convert_seed(&self, seed: Option<&str>) -> Result<Option<ed25519_sign::Seed>, CryptoError> {
-        trace!("convert_seed >>> seed: {:?}", seed);
+        trace!("convert_seed >>> seed: {:?}", secret!(seed));
 
         if seed.is_none() {
             trace!("convert_seed <<< res: None");
@@ -375,7 +405,7 @@ impl CryptoService {
 
         let res = ed25519_sign::Seed::from_slice(bytes.as_slice())?;
 
-        trace!("convert_seed <<< res: {:?}", res);
+        trace!("convert_seed <<< res: {:?}", secret!(&res));
 
         Ok(Some(res))
     }
@@ -436,7 +466,8 @@ mod tests {
     fn create_my_did_with_works_for_empty_info() {
         let service = CryptoService::new();
         let did_info = MyDidInfo { did: None, cid: None, seed: None, crypto_type: None };
-        service.create_my_did(&did_info).unwrap();
+        let my_did = service.create_my_did(&did_info);
+        assert!(my_did.is_ok());
     }
 
     #[test]
@@ -521,7 +552,8 @@ mod tests {
 
         let message = r#"message"#;
         let (_, my_key) = service.create_my_did(&did_info).unwrap();
-        service.sign(&my_key, message.as_bytes()).unwrap();
+        let sig = service.sign(&my_key, message.as_bytes());
+        assert!(sig.is_ok());
     }
 
     #[test]
@@ -587,7 +619,8 @@ mod tests {
         let (_, my_key) = service.create_my_did(&did_info).unwrap();
         let (their_did, _) = service.create_my_did(&did_info.clone()).unwrap();
         let their_did = Did::new(their_did.did, their_did.verkey);
-        service.encrypt(&my_key, &their_did.verkey, msg.as_bytes()).unwrap();
+        let encrypted_message = service.encrypt(&my_key, &their_did.verkey, msg.as_bytes());
+        assert!(encrypted_message.is_ok());
     }
 
     #[test]
@@ -653,7 +686,8 @@ mod tests {
         let did_info = MyDidInfo { did: None, cid: None, seed: None, crypto_type: None };
         let (did, _) = service.create_my_did(&did_info.clone()).unwrap();
         let did = Did::new(did.did, did.verkey);
-        service.encrypt_sealed(&did.verkey, msg.as_bytes()).unwrap();
+        let encrypted_message = service.encrypt_sealed(&did.verkey, msg.as_bytes());
+        assert!(encrypted_message.is_ok());
     }
 
     #[test]
@@ -666,5 +700,35 @@ mod tests {
         let encrypted_message = service.encrypt_sealed(&encrypt_did.verkey, msg).unwrap();
         let decrypted_message = service.decrypt_sealed(&key, &encrypted_message).unwrap();
         assert_eq!(msg, decrypted_message.as_slice());
+    }
+
+    #[test]
+    fn authenticated_encrypt_works() {
+        let service = CryptoService::new();
+        let msg = "some message".as_bytes();
+        let did_info = MyDidInfo { did: None, cid: None, seed: None, crypto_type: None };
+        let ( _ , my_key) = service.create_my_did(&did_info).unwrap();
+        let (their_did, _their_key) = service.create_my_did(&did_info.clone()).unwrap();
+        let their_did_for_encrypt = Did::new(their_did.did, their_did.verkey);
+        let encrypted_message = service.authenticated_encrypt(&my_key, &their_did_for_encrypt.verkey, msg);
+        assert!(encrypted_message.is_ok());
+    }
+
+    #[test]
+    fn authenticated_encrypt_and_authenticated_decrypt_works() {
+        let service = CryptoService::new();
+        let msg = "some message";
+        let did_info = MyDidInfo { did: None, cid: None, seed: None, crypto_type: None };
+        let (my_did, my_key) = service.create_my_did(&did_info).unwrap();
+        let my_key_for_encrypt = my_key.clone();
+        let their_did_for_decrypt = Did::new(my_did.did, my_did.verkey);
+        let (their_did, their_key) = service.create_my_did(&did_info.clone()).unwrap();
+        let my_key_for_decrypt = their_key.clone();
+        let their_did_for_encrypt = Did::new(their_did.did, their_did.verkey);
+        let encrypted_message = service.authenticated_encrypt(&my_key_for_encrypt, &their_did_for_encrypt.verkey, msg.as_bytes()).unwrap();
+        let (their_vk, decrypted_message) = service.authenticated_decrypt(&my_key_for_decrypt,&encrypted_message).unwrap();
+        assert_eq!(msg.as_bytes().to_vec(), decrypted_message);
+        assert_eq!(&their_vk, &their_did_for_decrypt.verkey);
+
     }
 }