+++ vcx/libvcx/src/utils/error.rs
@@ -98,6 +98,10 @@ pub static DID_ALREADY_EXISTS_IN_WALLET: Error = Error { code_num: 1083, message
 pub static DUPLICATE_MASTER_SECRET: Error = Error { code_num: 1084, message: "Attempted to add a Master Secret that already existed in wallet"};
 pub static THREAD_ERROR: Error = Error{ code_num: 1085, message: "Unable to create thread"};
 pub static INVALID_PROOF_REQUEST: Error = Error{ code_num: 1086, message: "Proof Request Passed into Libindy Call Was Invalid"};
+pub static MISSING_PAYMENT_METHOD: Error = Error{ code_num: 1087, message: "Configuration is missing the Payment Method parameter"};
+pub static DUPLICATE_SCHEMA: Error = Error{ code_num: 1088, message: "Duplicate Schema: Ledger Already Contains Schema For Given DID, Version, and Name Combination"};
+pub static UKNOWN_LIBINDY_TRANSACTION_REJECTION: Error = Error{ code_num: 1089, message: "Unknown Libindy Rejection"};
+pub static LOGGING_ERROR: Error = Error{ code_num: 1090, message: "Logging Error" };
 
 lazy_static! {
     static ref ERROR_C_MESSAGES: HashMap<u32, CString> = {
@@ -189,6 +193,11 @@ lazy_static! {
         insert_c_message(&mut m, &INVALID_LEDGER_RESPONSE);
         insert_c_message(&mut m, &THREAD_ERROR);
         insert_c_message(&mut m, &INVALID_PROOF_REQUEST);
+        insert_c_message(&mut m, &DUPLICATE_SCHEMA);
+        insert_c_message(&mut m, &UKNOWN_LIBINDY_TRANSACTION_REJECTION);
+        insert_c_message(&mut m, &MISSING_PAYMENT_METHOD);
+        insert_c_message(&mut m, &LOGGING_ERROR);
+
         m
     };
 }