+++ vcx/wrappers/python3/tests/conftest.py
@@ -1,14 +1,34 @@
 import pytest
+import time
 from vcx.api.vcx_init import vcx_init
+from vcx.common import shutdown as vcx_shutdown
+
+flag = False
 
 
 @pytest.mark.asyncio
 @pytest.fixture
 async def vcx_init_test_mode():
-    if not hasattr(vcx_init, "open"):
-        vcx_init.open = None
+    global flag
 
-    if not vcx_init.open:
+    if not flag:
         await vcx_init('ENABLE_TEST_MODE')
-        vcx_init.open = True
+        flag = True
 
+
+@pytest.fixture
+async def cleanup():
+
+    def _shutdown(erase):
+        global flag
+        vcx_shutdown(erase)
+        if flag:
+            flag = False
+
+    return _shutdown
+
+
+@pytest.fixture(scope='session', autouse=True)
+def wait_libindy():
+    yield
+    time.sleep(1) # FIXME IS-1060