+++ vcx/wrappers/python3/tests/test_wallet.py
@@ -1,6 +1,5 @@
 import pytest
 from vcx.error import VcxError, ErrorCode
-from vcx.common import shutdown
 from vcx.api.wallet import *
 import json
 
@@ -38,11 +37,13 @@ async def test_send_tokens():
     receipt = await Wallet.send_tokens(0,1,"address")
     assert receipt
 
+
 @pytest.mark.asyncio
 @pytest.mark.usefixtures('vcx_init_test_mode')
 async def test_create_payment_address():
     address = await Wallet.create_payment_address()
     assert address
+
 
 @pytest.mark.asyncio
 @pytest.mark.usefixtures('vcx_init_test_mode')
@@ -50,10 +51,12 @@ async def test_create_payment_address_with_seed():
     address = await Wallet.create_payment_address("0000000000000000000000WHATEVER00")
     assert address
 
+
 @pytest.mark.asyncio
 @pytest.mark.usefixtures('vcx_init_test_mode')
 async def test_validate_payment_address():
     await Wallet.validate_payment_address('sov:1:1234')
+
 
 @pytest.mark.asyncio
 @pytest.mark.usefixtures('vcx_init_test_mode')
@@ -73,6 +76,7 @@ async def test_wallet_storage():
     }
     assert (json.loads(await Wallet.get_record(TYPE, ID, OPTIONS)) == record)
 
+
 @pytest.mark.asyncio
 async def test_wallet_search():
     search_handle = await Wallet.open_search(TYPE, QUERY_JSON, None)
@@ -84,41 +88,42 @@ async def test_wallet_search():
     with pytest.raises(VcxError) as e:
         await Wallet.export("/tmp/output.wallet", "backupKey")
 
+
 @pytest.mark.asyncio
-async def test_import_wallet_failures():
+async def test_import_wallet_failures(vcx_init_test_mode, cleanup):
     with pytest.raises(VcxError) as e:
         await Wallet.import_wallet('Invalid Json')
     assert ErrorCode.InvalidJson == e.value.error_code
-    shutdown(True)
+    cleanup(True)
 
     config = {'wallet_name': '', 'wallet_key': '', 'exported_wallet_path': '', 'backup_key': ''}
     with pytest.raises(VcxError) as e:
         await Wallet.import_wallet(json.dumps(config))
     assert ErrorCode.IOError == e.value.error_code
-    shutdown(True)
+    cleanup(True)
 
     config = {'wallet_key': '', 'exported_wallet_path': '', 'backup_key': ''}
     with pytest.raises(VcxError) as e:
         await Wallet.import_wallet(json.dumps(config))
     assert ErrorCode.MissingWalletName == e.value.error_code
-    shutdown(True)
+    cleanup(True)
 
     config = {'wallet_name': '', 'exported_wallet_path': '', 'backup_key': ''}
     with pytest.raises(VcxError) as e:
         await Wallet.import_wallet(json.dumps(config))
     assert ErrorCode.MissingWalletKey == e.value.error_code
-    shutdown(True)
+    cleanup(True)
 
     config = {'wallet_name': '', 'wallet_key': '', 'backup_key': ''}
     with pytest.raises(VcxError) as e:
         await Wallet.import_wallet(json.dumps(config))
     assert ErrorCode.MissingExportedWalletPath == e.value.error_code
-    shutdown(True)
+    cleanup(True)
 
     config = {'wallet_name': '', 'wallet_key': '', 'exported_wallet_path': ''}
     with pytest.raises(VcxError) as e:
         await Wallet.import_wallet(json.dumps(config))
     assert ErrorCode.MissingBackupKey == e.value.error_code
-    shutdown(True)
+    cleanup(True)